+++ test/lbm/D3Q27.cpp
@@ -17,27 +17,27 @@ SCENARIO("D3Q27 velocity set properties")
 
         std::move_only_function<std::size_t()> model_dimension = &Set::dimension;
         std::move_only_function<std::size_t()> model_size = &Set::size;
-        std::move_only_function<std::array<std::array<int, Set::dimension()>, Set::size()>()>
+        std::move_only_function<std::array<std::array<float, Set::dimension()>, Set::size()>()>
             model_velocities = &Set::velocities;
         std::move_only_function<std::array<float, Set::size()>()> model_weights = &Set::weights;
 
         const auto expectedDimension{3};
         const auto expectedSize{27};
-        const auto expectedVelocities = std::array<std::array<int, 3>, 27>{
-            std::array<int, 3>{0, 0, 0},   std::array<int, 3>{1, 0, 0},
-            std::array<int, 3>{-1, 0, 0},  std::array<int, 3>{0, 1, 0},
-            std::array<int, 3>{0, -1, 0},  std::array<int, 3>{0, 0, 1},
-            std::array<int, 3>{0, 0, -1},  std::array<int, 3>{1, 1, 0},
-            std::array<int, 3>{-1, 1, 0},  std::array<int, 3>{1, -1, 0},
-            std::array<int, 3>{-1, -1, 0}, std::array<int, 3>{1, 0, 1},
-            std::array<int, 3>{-1, 0, 1},  std::array<int, 3>{1, 0, -1},
-            std::array<int, 3>{-1, 0, -1}, std::array<int, 3>{0, 1, 1},
-            std::array<int, 3>{0, -1, 1},  std::array<int, 3>{0, 1, -1},
-            std::array<int, 3>{0, -1, -1}, std::array<int, 3>{1, 1, 1},
-            std::array<int, 3>{-1, 1, 1},  std::array<int, 3>{1, -1, 1},
-            std::array<int, 3>{-1, -1, 1}, std::array<int, 3>{1, 1, -1},
-            std::array<int, 3>{-1, 1, -1}, std::array<int, 3>{1, -1, -1},
-            std::array<int, 3>{-1, -1, -1}
+        const auto expectedVelocities = std::array<std::array<float, 3>, 27>{
+            std::array<float, 3>{0, 0, 0},   std::array<float, 3>{1, 0, 0},
+            std::array<float, 3>{-1, 0, 0},  std::array<float, 3>{0, 1, 0},
+            std::array<float, 3>{0, -1, 0},  std::array<float, 3>{0, 0, 1},
+            std::array<float, 3>{0, 0, -1},  std::array<float, 3>{1, 1, 0},
+            std::array<float, 3>{-1, 1, 0},  std::array<float, 3>{1, -1, 0},
+            std::array<float, 3>{-1, -1, 0}, std::array<float, 3>{1, 0, 1},
+            std::array<float, 3>{-1, 0, 1},  std::array<float, 3>{1, 0, -1},
+            std::array<float, 3>{-1, 0, -1}, std::array<float, 3>{0, 1, 1},
+            std::array<float, 3>{0, -1, 1},  std::array<float, 3>{0, 1, -1},
+            std::array<float, 3>{0, -1, -1}, std::array<float, 3>{1, 1, 1},
+            std::array<float, 3>{-1, 1, 1},  std::array<float, 3>{1, -1, 1},
+            std::array<float, 3>{-1, -1, 1}, std::array<float, 3>{1, 1, -1},
+            std::array<float, 3>{-1, 1, -1}, std::array<float, 3>{1, -1, -1},
+            std::array<float, 3>{-1, -1, -1}
         };
         const auto expectedWeights = std::array<float, 27>{
             8.0F / 27,  2.0F / 27,  2.0F / 27,  2.0F / 27,  2.0F / 27,  2.0F / 27,  2.0F / 27,
@@ -78,7 +78,7 @@ SCENARIO("D3Q27 velocity set properties")
 
             THEN("Sum of velocities is zero")
             {
-                std::array<int, 3> sum{0, 0, 0};
+                std::array<float, 3> sum{0, 0, 0};
                 for (const auto& velocity : velocities)
                 {
                     sum[0] += velocity[0];
@@ -86,7 +86,7 @@ SCENARIO("D3Q27 velocity set properties")
                     sum[2] += velocity[2];
                 }
 
-                CHECK((sum == std::array<int, 3>{0, 0, 0}));
+                CHECK((sum == std::array<float, 3>{0, 0, 0}));
             }
         }
